requirements.txt
python-pptx
pandas
numpy
openpyxl